Restructured, but little code changes
