+++ travel_app/static/modules/tripBuilder/modal/modal-template.html
