+++ travel_app/static/modules/tripBuilder/rentals/rentals-template.html
@@ -1,0 +1,5 @@
+<div class="contentArea" id="rentals">
+    <tab-content params="value: rentals">
+
+      </tab-content>
+</div>
