+++ travel_app/static/modules/utils/loader.js
