+++ travel_app/static/modules/utils/utilFunctions.js
@@ -1,0 +1,19 @@
+export {logging, formatDate}
+
+function logging(...args) {
+    console.log(...args)
+}
+
+function formatDate(date) {
+    var d = new Date(date),
+        month = '' + (d.getMonth() + 1),
+        day = '' + d.getDate(),
+        year = d.getFullYear();
+
+    if (month.length < 2) 
+        month = '0' + month;
+    if (day.length < 2) 
+        day = '0' + day;
+
+    return [year, month, day].join('-');
+}
